pyproject.toml
[tool.poetry]
name = "rlt"
version = "0.1.0"
description = ""
authors = ["sanek <alexandr.panteleev2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
motor = "^3.3.1"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
